# [wip] trie: range verification based on stacktrie