--- a/trie/stackproof.go
+++ b/trie/stackproof.go
@@ -397,19 +397,14 @@ func VerifyRangeProofWithStack(rootHash common.Hash, firstKey []byte, keys [][]b
 
 // wrarpWriteFunction returns a NodeWriteFunc which filters away writes that are
 // on the boundary: parents of first/last.
-func wrapWriteFunction(origin, first, last []byte, w NodeWriteFunc) NodeWriteFunc {
+func wrapWriteFunction(origin, last []byte, w NodeWriteFunc) NodeWriteFunc {
 	if w == nil {
 		return nil
 	}
 	var originBorder = keybytesToHex(origin)
-	var leftBorder = keybytesToHex(first)
 	var rightBorder = keybytesToHex(last)
-	return func(origin common.Hash, path []byte, hash common.Hash, blob []byte) {
+	return func(path []byte, hash common.Hash, blob []byte) {
 		if bytes.HasPrefix(originBorder, path) {
-			//fmt.Printf("path %x  tainted left (parent to %x)\n", path, leftBorder)
-			return
-		}
-		if bytes.HasPrefix(leftBorder, path) {
 			//fmt.Printf("path %x  tainted left (parent to %x)\n", path, leftBorder)
 			return
 		}
@@ -417,6 +412,6 @@ func wrapWriteFunction(origin, first, last []byte, w NodeWriteFunc) NodeWriteFun
 			//fmt.Printf("path %x  tainted right (parent to %x)\n", path, rightBorder)
 			return
 		}
-		w(origin, path, hash, blob)
+		w(path, hash, blob)
 	}
 }