… eth/protocols/snap/sync_test.go
… trie/proof_test.go
… trie/stackproof_test.go
… trie/stacktrie.go
… trie/trie_test.go