This PR implements 

- Ability to 'shape' a stacktrie based on a proof, to lock in the "prehistory" or "left side" of a trie. 
- Abilty to fill a stacktrie with hashes based on a proof, in order to complete the "right- side" with the hashes in a proof. 
(This operation does not mutate the shape of the trie)
- Combines these two abilities into a new range-verifier. 

```
name                    old time/op    new time/op    delta
VerifyRangeProof10-8      72.0µs ±12%    82.9µs ±11%  +15.19%  (p=0.016 n=5+5)
VerifyRangeProof100-8      594µs ±12%     284µs ±11%  -52.17%  (p=0.008 n=5+5)
VerifyRangeProof1000-8    5.24ms ±14%    2.68ms ± 5%  -48.88%  (p=0.008 n=5+5)
VerifyRangeProof5000-8    18.3ms ± 7%    12.4ms ± 9%  -31.99%  (p=0.008 n=5+5)
VerifyRangeProof10K-8     30.3ms ±15%    25.2ms ± 2%  -17.10%  (p=0.008 n=5+5)

name                    old alloc/op   new alloc/op   delta
VerifyRangeProof10-8      36.1kB ± 1%    27.9kB ± 9%  -22.56%  (p=0.016 n=4+5)
VerifyRangeProof100-8      183kB ± 3%      40kB ± 5%  -78.10%  (p=0.008 n=5+5)
VerifyRangeProof1000-8    1.83MB ± 1%    0.22MB ± 2%  -88.27%  (p=0.008 n=5+5)
VerifyRangeProof5000-8    10.4MB ± 0%     1.0MB ± 1%  -89.93%  (p=0.008 n=5+5)
VerifyRangeProof10K-8     21.4MB ± 0%     2.2MB ± 1%  -89.75%  (p=0.008 n=5+5)

name                    old allocs/op  new allocs/op  delta
VerifyRangeProof10-8         432 ± 0%       542 ± 4%  +25.48%  (p=0.016 n=4+5)
VerifyRangeProof100-8      1.77k ± 2%     0.88k ± 4%  -50.24%  (p=0.008 n=5+5)
VerifyRangeProof1000-8     16.2k ± 1%      6.0k ± 1%  -62.71%  (p=0.008 n=5+5)
VerifyRangeProof5000-8     85.3k ± 0%     30.1k ± 1%  -64.72%  (p=0.008 n=5+5)
VerifyRangeProof10K-8       172k ± 0%       61k ± 1%  -64.39%  (p=0.008 n=5+5)
```
The new range-verifier is not necessarly a whole lot faster, but is a lot lower in memory usage -- particularly when handling larger slices (our snapsync handles around 10K items per request). 

This PR builds upon a few other PRs, and is thus still draft.