@holiman 2023-10-09
The integration is still not quite 100%

```
Oct 09 19:12:33 bench07.ethdevops.io geth INFO [10-09|17:12:33.025] Syncing: chain download in progress synced=99.96% chain=585.02GiB headers=18,314,323@8.41GiB bodies=18,306,383@414.21GiB receipts=18,306,383@162.40GiB eta=11.403s
Oct 09 19:12:41 bench07.ethdevops.io geth INFO [10-09|17:12:41.662] Syncing: chain download in progress synced=99.97% chain=585.28GiB headers=18,314,323@8.41GiB bodies=18,308,726@414.43GiB receipts=18,308,726@162.44GiB eta=8.039s
Oct 09 19:12:42 bench07.ethdevops.io factor INFO [10-09|17:12:42.027] New head block number=18,314,324 hash=12856b..9822b6
Oct 09 19:12:42 bench07.ethdevops.io geth INFO [10-09|17:12:42.045] Forkchoice requested sync to new head number=18,314,324 hash=12856b..9822b6 finalized=18,314,235
Oct 09 19:12:49 bench07.ethdevops.io geth INFO [10-09|17:12:49.799] Syncing: chain download in progress synced=99.98% chain=585.53GiB headers=18,314,324@8.41GiB bodies=18,311,270@414.64GiB receipts=18,311,270@162.48GiB eta=4.387s
Oct 09 19:12:52 bench07.ethdevops.io factor INFO [10-09|17:12:52.606] New head block number=18,314,325 hash=fa2a00..fd174e
Oct 09 19:12:52 bench07.ethdevops.io geth INFO [10-09|17:12:52.628] Forkchoice requested sync to new head    number=18,314,325 hash=fa2a00..fd174e finalized=18,314,235
Oct 09 19:12:58 bench07.ethdevops.io geth INFO [10-09|17:12:58.236] Syncing: chain download in progress synced=100.00% chain=585.83GiB headers=18,314,325@8.41GiB bodies=18,313,550@414.90GiB receipts=18,313,550@162.52GiB eta=1.113s
Oct 09 19:13:00 bench07.ethdevops.io geth INFO [10-09|17:13:00.520] Rebuilding state snapshot
Oct 09 19:13:00 bench07.ethdevops.io geth INFO [10-09|17:13:00.522] Committed new head block number=18,314,213 hash=4abb53..3ff808
Oct 09 19:13:00 bench07.ethdevops.io geth INFO [10-09|17:13:00.522] Resuming state snapshot generation root=4283b0..8547a9 accounts=0 slots=0 storage=0.00B dangling=0 elapsed=1.544ms
Oct 09 19:13:00 bench07.ethdevops.io geth WARN [10-09|17:13:00.853] Trie prefetcher failed opening trie      root=448d3f..4e590b err="missing trie node 448d3f8cdf8ff81b87b310310b8923a809a3249eb9d8ba05dbe8c83b2e4e590b (owner 7eb6a6f61feceb1bb0d0e48cb8880be2dbbe987875f9e41d5f8e5dc91cac78a1) (path ) <nil>"
Oct 09 19:13:01 bench07.ethdevops.io geth WARN [10-09|17:13:00.953] Failed to load old bad blocks error="pebble: not found"
Oct 09 19:13:01 bench07.ethdevops.io geth ERROR[10-09|17:13:00.957]
Oct 09 19:13:01 bench07.ethdevops.io geth ########## BAD BLOCK #########
Oct 09 19:13:01 bench07.ethdevops.io geth Block: 18314214 (0x34272d45968be1aa458680d34001759c62378245853a1495b11ea4a0d8d8cb20)
Oct 09 19:13:01 bench07.ethdevops.io geth Error: invalid merkle root (remote: a8db6a2db0f4c8364db7140a1c60f13f59c7e2892f720139c642dafdc00904b1 local: 843061dd862a8eda1d6d26bfc09143a8ef87d2dbb380b0bfeeac89321c52f567) dberr: missing trie node 448d3f8cdf8ff81b87b310310b8923a809a3249eb9d8ba05dbe8c83b2e4e590b (owner 7eb6a6f61feceb1bb0d0e48cb8880be2dbbe987875f9e41d5f8e5dc91cac78a1) (path ) <nil>
Oct 09 19:13:01 bench07.ethdevops.io geth Platform: geth (devel) go1.21.2 amd64 linux
Oct 09 19:13:01 bench07.ethdevops.io geth VCS: 200336c5-20231009
Oct 09 19:13:01 bench07.ethdevops.io geth Chain config: &params.ChainConfig{ChainID:1, HomesteadBlock:1150000, DAOForkBlock:1920000, DAOForkSupport:true, EIP150Block:2463000, EIP155Block:2675000, EIP158Block:2675000, ByzantiumBlock:4370000, ConstantinopleBlock:7280000, PetersburgBlock:7280000, IstanbulBlock:9069000, MuirGlacierBlock:9200000, BerlinBlock:12244000, LondonBlock:12965000, ArrowGlacierBlock:13773000, GrayGlacierBlock:15050000, MergeNetsplitBlock:<nil>, ShanghaiTime:(*uint64)(0xc0003ce880), CancunTime:(*uint64)(nil), PragueTime:(*uint64)(nil), VerkleTime:(*uint64)(nil), TerminalTotalDifficulty:58750000000000000000000, TerminalTotalDifficultyPassed:true, Ethash:(*params.EthashConfig)(0x2d04cc0), Clique:(*params.CliqueConfig)(nil), IsDevMode:false}
Oct 09 19:13:01 bench07.ethdevops.io geth Receipts: 
```