@holiman 2024-04-27
This worked well, but when I benchmarked it, the effect of the lowered mem-usage during snap sync was hardly noticeable. 
The PR has bitrotted quite a bit, so I'll close it now, but it migth be worth revisiting if we ever want to have the ability to have "fully correct partial stacktries" (proof-initialized, proof-ended, resulting in correct root and committed nodes) , instead of what we have now, "mostly correct partial stacktries" (where we need to discard all edge-elements and get an incorrect root)